Added main-loop delay because my test box runs too fast
This is a tricky one to get right, so I'm going to keep fiddling

+++ src/physics_bullet.cpp
@@ -218,7 +218,7 @@ void PhysicsBullet::delRigidBody(btRigidBody* body)
 **/
 void PhysicsBullet::stepTime(int ms)
 {
-	dynamicsWorld->stepSimulation(ms / 1000.0f, 10);
+	dynamicsWorld->stepSimulation( ((float)ms) / 1000.0f , 10);
 }
